[refactor] ChatController 패키지 이동 및 서비스 메소드들 네임 변경
